modify: 더미데이터/타입, 분리
